+++ components/admin/works-management/DeleteWork.jsx
@@ -6,29 +6,35 @@ import React from 'react'
 import { FaTrashAlt } from 'react-icons/fa'
 
 const DeleteWork = (workInfo) => {
+  const [isConnected, setIsConnected] = useState()
+  useEffect(() => {
+    setIsConnected(localStorage.getItem("token"))
+  })
 
   const router = useRouter()
   
   const workId = Object.values(workInfo)[0]
 
   const handleWorkDeletion = async () => {
-    const confirmDeletion = confirm("Êtes vous sûr.e de vouloir supprimer ce projet ?")
-    
-    if(confirmDeletion){
-      const res = await fetch("api/works/" + workId, {
-        method: "DELETE",
-        headers: {
-          "Content-type": "application/json",
-        },
-      })
+    if(isConnected){
+      const confirmDeletion = confirm("Êtes vous sûr.e de vouloir supprimer ce projet ?")
+      
+      if(confirmDeletion){
+        const res = await fetch("api/works/" + workId, {
+          method: "DELETE",
+          headers: {
+            "Content-type": "application/json",
+          },
+        })
 
-      if(res.status === 200){
-        alert("Work deleted!") 
-        router.push("/admin")
-      } else {
-        alert("Something went wrong while creating the project.")
+        if(res.status === 200){
+          alert("Work deleted!") 
+          router.push("/admin")
+        } else {
+          alert("Something went wrong while creating the project.")
+        }
       }
-    };
+    }
   }
 
   return (
